# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"RepositoryDTO_unknown_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {},
					"data": {}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteModel": {
				"properties": {
					"ids": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"ids"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RepositoryDTO_any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {},
					"data": {}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TypeSort": {
				"enum": [
					0,
					1
				],
				"type": "number"
			},
...
			"IngredientModel": {
				"properties": {
					"name": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"imageUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPagination_any_": {
				"properties": {
					"records": {
						"items": {},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"records",
					"total",
					"page",
					"pageSize",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RepositoryDTO_IPagination_any__": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {},
					"data": {
						"$ref": "#/components/schemas/IPagination_any_"
					}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"JWT": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": ""
	},
	"paths": {
...
		"/ingredient": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "trả về một danh sách món ăn",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredient"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredient"
				],
				"security": [
					{
						"JWT": [
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IngredientModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredient"
				],
				"security": [
					{
						"JWT": [
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
...
		"/hint/ingredient": {
			"get": {
				"operationId": "GetHintIngredient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_IPagination_any__"
								}
							}
						}
					}
				},
				"description": "lọc theo nguyên liệu và sợ thích của người dùng",
				"tags": [
					"Hint Recipe"
				],
				"security": [
					{
						"JWT": [
							null,
							null
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "ingredientIds",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/hint/ingredient/public": {
			"get": {
				"operationId": "GetHintIngredientPublic",
				"responses": {
					"200": {
						"description": "Tìm kiếm theo nguyên liệu như theo kiểu public ai sài cũng đc dùng co tài khoản gust",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_IPagination_any__"
								}
							}
						}
					}
				},
				"tags": [
					"Hint Recipe"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "ingredientIds",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/hint/recipeUser": {
			"get": {
				"operationId": "GetHintUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_IPagination_any__"
								}
							}
						}
					}
				},
				"tags": [
					"Hint Recipe"
				],
				"security": [
					{
						"JWT": [
							null,
							null
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/hint/recipeSimilar/{id}": {
			"get": {
				"operationId": "GetHintSimilar",
				"responses": {
					"200": {
						"description": "Lấy món ăn tương tự với id trên",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_IPagination_any__"
								}
							}
						}
					}
				},
				"tags": [
					"Hint Recipe"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						},
						"example": 1
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/hint/recipeHightly": {
			"get": {
				"operationId": "GetHighlyAppreciated",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_IPagination_any__"
								}
							}
						}
					}
				},
				"tags": [
					"Hint Recipe"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}